Handle empty nodes and convertable types to match iOS impl behavior (#24)
I found that Android reanimated code that runs on iOS would often crash because of nodes being used and not registered. This seems to be a quite frequent usecase that some nodes are created and used e.g. in "set" statement but because of the way views are configured they may not be attached to the view hierarchy and hence missing from native registry. 
Obj-c handles that case gracefully as all methods can be called on nullable objects. This is not the case in Java, and this PR attempts to mitigate that issue.