Change drop procedure to rely on events instead of callbacks.
Here we change the way we handled nodes dropped on the native side. Since we don't know at the point of ropping a native node if it is going to be kept on the javascript side and reused later on, we have to copy the last seen value on the native side to be used later in case the node gets recreated on the native side. Previously we've been using getValue call. But because of an excessive amount of RN callback's that this way would produce in some cases we had to switch to using events instead (RN 61 added a check for a large number of callbacks being registered).