Fixes transitioning API delay when delayMs is not set (#335)
After implamenting "Slow Transitions" feature an issue was introduced with Transitioning API as you can see here tomasgcs@4ef8b81#diff-a7ea5763fc80cde0b741dca6a5095a18R46

Previously beginTime variable was not set at all which was equal to 0, but after this specific line was added the value is different, and it causes animations to delay even if delayMs value is not set, this happens only when animation is is being executed multiple times.

Not sure why this is happening but most likely that CACurrentMediaTime() introduces delay even if it takes current time. In my case it is very important to run animations as fast as possible so I've sumbitted a fix here.