Fix Android Gradle setup (#463)
Trying to open or build the android/ folder was not possible, due to a missing buildscript block:

ERROR: Plugin with id 'com.android.library' not found.
After adding the reference to the source of the library plugin, the following error showed up:

ERROR: Failed to resolve: androidx.transition:transition:1.1.0
This was caused by missing entries in the repositories block. Added the necessary references to resolve the errors.

It is now possible to individually open and build the android/ folder in Android Studio (or build it on command line, after generating a Gradle wrapper).

Note that these changes will not affect consumer (app) projects - The Android Gradle plugin is included conditionally, meaning the dependency classpath only gets added when the project is opened standalone. This is better than simply removing it (as seen in #148)