Fix issue with some events not propagating correctly via Animated.event (#531)
The issue has been discovered while testing onScroll listener with react-native-webview. After investigating it turned out that there are two problems:
1) WebView was expecting a listener to be assigned to the native component and only with the listener preset the events would be dispatched
2) On iOS we weren't translating event names properly. Some older component still follow "topSth" naming convention while we were solely relying on "onSth" convention for direct event.

In order to address (1) we now assign an empty listener to component properties that have Animated.event object assigned. This way native component knows there is a listener assigned and can start dispatching events that we hijack on the native side.

To fix (2) we matched the orifinal Animated implementation and adapted RCTNormalizeInputEventName utility method from RN core that normalizes event names (it actually changes all names to "topSth" convetion but at least it is now consistent across the board).