[web] Changed how values are updated with Animated.event to prev… (#536)
the purpose of this PR is to remove the non-standard method call in favor of an API that has more parity with native. This should also fix some issues where gesture updates freeze.
This change won't be backwards compatible and will require a version of RNGH with this software-mansion/react-native-gesture-handler#905. However RNGH will be fully backwards compatible with reanimated. This shouldn't be a huge issue as reanimated barely works on the web currently.