fix: Transitionning.View in mock.js
Issue: Previously `Transitioning.View` in the mock used the `createMockComponent`. The issue is that this function doesn't return a component which has the `animateNextTransition()` function. So when doing `ref.current.animateNextTransition()` in your tests, this line will crash because of the function not existing.

Fix: This commit ensures that `Transitioning.View` is mocked by using `createTransitioningComponent` view function which provides the `animateNextTransition` to the ref.