Fix misspelling in installation.md
Based on the [playground diff](https://github.com/software-mansion-labs/reanimated-2-playground/commit/f6f2b77496bc00601150f98ea19a341f844d06a3), AppDelegate.m should be changed to *.mm filename and hence renamed to AppDelegate.mm.  But the docs seem to add a second change, which is to use "a" instead of "e" in the otherwise correct spelling of "delegate" and hence AppDelagate.mm.

Figured that this was an unintentional misspelling that may lead to some small but unnecessary confusion and worth fixing.