Enable starting animations from RN thread. (#904)
This change makes it possible to start animations using `withTiming` and similar methods directly on the RN JS thread.

For that to work the following changes were made:
1) On the c++ side, when updating mutable value we now need to call setter. As the value is kept on the "remote" side we run the setter on the UI thread using scheduler
2) On the JS side we need to distinguish between cases when `withTiming` and similar methods are called to determine initial style in `useAnimatedStyle` vs when they are called to be set under `.value` prop. To do that we wrap initial updater calls in `useAnimatedStyle` with a method that sets updates global variable, then read that variable and depending on its state we either return the initial value or the animation.
3) All animation definitions needed to be refactored. We now wrap all animations with `defineAnimation` call, that is responsible for doing the check decsribed in (2). On top of that `defineAnimation` wraps the animation definition such that we no longer return animation object but a worklet which generates that object. This is needed because if the animation object was constructed on the RN JS side, it'd be frozen and UI thred wouldn't be able to modify it while running animations. By passing a worklet we can initialize animation object on the UI side.

Because of the new approach for defining animations, the value setter now must consider two separate cases: when it gets animation object, or when it gets a worklet that outputs animation. We allow for those two options to be returned, otherwise we'd need to make adjustments in the useAnimatedStyle updater which are far more complex.

Note that the changes made to animation functions described in (3) are of a smaller scope that it may seem on the diff. This is because after wrapping animations in `defineAnimation` method we had to indent all the existing code. In `delay` and `loop` methods we transform variables passed to those methods to keep the changeset shorter.

#### Testing

To test this change I modified `AnimatedStyleUpdateExample` to use `withTiming`, `withSpring`, `delay` and `loop` directly in `onPress` handler.