[v1] Forbid using null in Animated.Value (#970)
Reanimated doesn't have any use case for holding `null` value in the `Animated.Value` besides using `defined` node to set it to some initial state, so we're relying on `undefined` being treated like that "empty" value.
In order to prevent some obscure errors when `null` is used we introduce explicit check for creating Value with `null` - it works when instantiating new Value and using imperative `value.setValue()`.

We don't need to change the `set` node, because it expects `double | AnimatedValue` to be used as it's the second argument.