Throw an error when trying to use new `interpolate` with nodes (#861)
## Description

I've found out I'm using it many times in my codebase (and supportive libraries) and thus it became difficult for me to get it renamed automatically (bc there were different imports' names etc). So I've added an error thrown while calling `interpolate` with a node (duck-typed).

## Changes

Throwing an error in the first argument seems to be a reanimated node. 
Also, I observer that currently it's not possible to use new features with default imports. I've added it because we were doing it previously.