Fix setting values inside procs
Current implementation of procs create proxes for every parameter used (which can be any animated node).
Those proxies hold a ref to the params as nodes, forwarding every method which can be used on them (`setValue()` for Values, `start()` and `stop()` for clocks, etc). When proc is called, a context for every param is created (which allows using procs in procs (in particular creating recursive functions)). Those contexts are used for memoization to ensure we don't reevaluate nodes that were calculated previously. The problem was in updating this memoized value when using `AnimatedParam.setValue()` - it only updated value of underlying node and left param value unchanged. It's updated in current call ID in opposition to Value which is updated in previous call ID. This ensures that nodes which rely on Value will be updated appropriately. Only AnimatedCallFunction depends on AnimatedParam, so we can update it in the current context which will ensure that next node evaluated in function will use updated value of AnimatedValue.