Fix color processing on the JS thread (#1055)
Thi fixes a problem with processColor method crashing when executed on the JS thread.

The reason for the crash was that the cached object containing matches wasn't initialized on the JS thread but only on the UI thread. We used RemoteObject for that purpose, as we needed to make sure that RegEx instances we put in that object are constructed in the right VM (those objects cannot be simply moved between UI and JS). In order for matchers to be available on the JS thread, we also need to have them instantiated on the JS thread.

This PR changes the matchers caching logic such that we select the correct cache with matchers depending on the thread we run in, and we also instantiate the matchers on both JS and UI threads in order to cache them in the correct cache.