Fix Navigation crash (#1187)
Fixing: https://github.com/software-mansion/react-native-reanimated/issues/1181
The problem linked above caused crashes in random places and after 1.5d of debugging I came up with an idea that there must be sth wrong with the memory management. Because of the problem could be reproduced with just one mapper and one animation it was clear for me that the root cause of the problem is in the notification of mapper dirtiness. We may want to notify the mapper which has been unregistered. In order to fix the problem, I added a code that removes listeners of unregistered mappers.