Fast refresh for gestures (#1165)
## Motivation

Motivated by [this issue](https://github.com/software-mansion/react-native-reanimated/issues/1155) this PR has been created.

## Description

To keep the hooks consistent `useAnimatedGestureHandler` is about to rebuild worklets on certain changes in the same manner as `useAnimatedStyle` or `useDerivedValue`. The decision whether to rebuild or not will depend on the dependency array which may be passed to the hook or not. If it is passed it is read directly from the argument, if not, it pulls the dependencies from the worklets' closure. Also in both cases the dependency array will be decorated with the worklets' hashes which means on every worklet's code change the rebuild will be triggered.

This is in fact a continuation of the works regarding fast refresh issues.

## Changes

`Hooks.js` - build dependencies for `useAnimatedGestureHandler` and rebuild worklet in `useEvent`
`createAnimatedComponent.js` - added logic of reattaching for event worklets
`FastRefreshTest.js` - added tests for the `useAnimatedGestureHandler` hook.