Add unit tests for the babel plugin (#1207)
## Description

The groundwork for addressing things like #823

I wanted to make changes to the babel plugin, but I couldn't see any focused tests for the plugin itself. From the look of things, we're relying on the test app here.

Since I don't want to break things, here are some fairly basic tests to cover the babel plugin.