Fix reattaching (#1217)
## Description

The following problem has been reported: the events are being reattached even when the handlers are not being rebuilt(which basically means reattaching is pointless). That resulted in a significant slowdown. The solution in this PR is to reattach events only for those handlers which body changed(basing on dependencies passed(or not) to the hook).

## Changes

`createAnimatedComponent.js` - reattach conditionally
`WorkletEventHandler.js` - add `reattachNeeded` flag