Shared value rerender upgrade (#1351)
## Description

This PR makes `useSharedValue` more consistent with other hooks when needed. Now it would change on every rerender when the value passed to the hook changes.
So in the following scenario
```
const [state, setState] = useState(0);
const sv = useSharedValue(state);
```
every time `state` changes(and component rerenders) `sv`'s value is updated to what's in `state`.

This is inconsistent with for instance `useState`, for which the state has to be set explicitly to get modified. However the default behavior persisted for backward compatibility.


Fixes https://github.com/software-mansion/react-native-reanimated/issues/1293