Remove ref from useAnimatedReaction hook (#1412)
## Description

This PR is a response for the problem with `useAnimatedReaction` hook. The problem is with using reference inside of that hook, because it persists the mapper's input values between rerenders. This may even work if we don't change the type of the input but the problem occurred when the input changed from `undefined` to `number`.

Reproducing code is in the issue.

Fixes https://github.com/software-mansion/react-native-reanimated/issues/1399