Handle worklet errors on hermes (#1491)
## Description

This PR refers to a situation in which there is a javascript error in a worklet that's being executed by jsi.
On iOS with JSC, it works as expected which means it throws a `std::exception` from jsi's `call` invocation.
Unfortunately on Android, it does not. This c++ code is being scheduled on the UI thread by the scheduler and the error that appears in `Scheduler.java` is: `com.facebook.JNI.UnknownCppException: Unknown`.
This comes with two problems:
1. Our example app crashes with `A/libc: Fatal signal 6 (SIGABRT), code -6`
This may be a problem with how we build the example app(As @tmikov suggested maybe e.g. linking problem)
2. When reanimated is installed in any other app it shows Redbox with an error message: `Unknown` - this gives almost no information on what went wrong.

The good news is that some error is thrown in c++ and can be caught. The bad news is that it's not a `std::exception` and we don't know its type so we cannot really get the proper message. Unless it changes we only can rethrow an error with a hardcoded message which indicates that something went wrong in a worklet. That's what is done in this PR.

Edit: Using the babel plugin I was able to fetch the path of the file containing the worklet in which the error occurred so it's now attached to the hardcoded error message.

## Changes

Just added catch clauses that catch anything for try blocks that include worklet execution.

I also repaired access to a module when using the `errorHandler`.

## Testing code

<details>
<summary>code</summary>

```
import { runOnUI } from 'react-native-reanimated';
import { View, Text, Button } from 'react-native';
import React, { useState } from 'react';

const Bug1 = () => {
  const val = {};

  runOnUI(() => {
    'worklet';
    val.f(4); // will call https://github.com/software-mansion/react-native-reanimated/blob/master/Common/cpp/SharedItems/ShareableValue.cpp#L347
  })();

  return <></>;
};

const Bug2 = () => {
  const wrklt = (x) => {
    'worklet';
    x.f();
  };

  runOnUI(() => {
    'worklet';
    wrklt(9); // will call https://github.com/software-mansion/react-native-reanimated/blob/master/Common/cpp/SharedItems/ShareableValue.cpp#L298
  })();

  return <></>;
};

export default function AnimatedStyleUpdateExample() {
  const [bug, setBug] = useState(0);

  const renderBug = () => {
    if (bug === 1) {
      return <Bug1 />;
    }
    if (bug === 2) {
      return <Bug2 />;
    }
  };

  return (
    <View>
      <Text>Pick a bug:</Text>
      <Button
        title="bug 1"
        onPress={() => {
          setBug(1);
        }}
      />
      <Button
        title="bug 2"
        onPress={() => {
          setBug(2);
        }}
      />
      {renderBug()}
    </View>
  );
}

```

</details>

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes