Add possibility to start layout animation when entering one hasn't finished (#2135)
Fixes: https://github.com/software-mansion/react-native-reanimated/issues/2119
When we get notified by ReactBatchObserver about the dirtiness of a component then we have to choose how should we animate that layout change. For instance, we have to decide if it should be entering, layout, or exiting animation.
Each component is treated as a state machine with the following lifecycle:
inactive -> appearing -> layout -> disappearing -> toRemove
Currently, we do not start a layout animation if the component is still entering. The problem with that is sometimes the coordinates of the component can change and our component will not be rendered properly.  You can read more about it in the issue linked above. The solution for that is simply to start a layout animation when the component is dirty. The problem with that solution is the sometimes component can be dirty which doesn't mean that the layout of the component has changed. For instance, the component can be marked dirty because one of its descendants had changed its layout.
To fix the solution I also added check if the startingValues (layout snapshot took as the first operation of UIManager's UI queue operations) are the same as targetValues ((layout snapshot took as the last operation of UIManager's UI queue operations). If they are it means that there was no UI operation that changed the layout of the component and it means that the dirtiness of the component is a false positive.  

I also added 3 examples that helped me test this pull-request.