Fix problems with polyfills (#2161)
## Description

Babel was changing some construction to non-workletized functions inside the worklet function. For example, `Object.assign()` was sometimes changed to babel implementation - this implementation wasn't worklet. That was the cause of the worklet runtime error.

Fixes https://github.com/software-mansion/react-native-reanimated/issues/2142

Known problem:
In a scenario when we have nested worklets the inner one won't be workletized with asString property of the outer one. It may cause problems in the future when sending functions between several threads will be more common. However, I don't think we have such examples in the reanimated world right now.