Fix build-script (#2355)
## Description

We have many dependencies to React Native, but react is changing. If we want to support many versions of RN we need to add some specific changes dedicated to react version. In Java, it is harder than C++ because we haven't `#ifdef` mechanism. So I added a patch-like mechanism to replace some specific files during compilation time for specific RN versions.