Fix problem with overlapping views (#2438)
## Description

Fixes: https://github.com/software-mansion/react-native-reanimated/issues/2432
Currently, we schedule leftover elements removal to avoid traversing the whole view tree every time onViewRemoval is called. However, with the current approach, it's possible that we remove elements on the next frame that should be removed instantly. To fix the problem I added an additional synchronous call of the removedLeftovers method as the last operation of a UI operation block.