Added support for shareable jsi objects and functions
When implementing jsi host objects or host functions that wraps native objects/funcs it would be nice to be able to share these between javascript runtimes.

This commit adds support for marking jsi host objects/host functions as shareable and handle them in a controlled manner.

It will always be the implementer's responsibility to ensure thread safety when using objects/funcs in different runtimes.

To mark a host object or host function as shareable, either call the ShareableValue::markHostAsShareable on the object/func, or make sure the host object returns true when the `get` method is called for the property ShareableValue::HOST_SHAREABLE.

- Added two new ValueTypes to support shareable host objects/functions.
- Added value wrappers for the new ValueTypes
- Added conversion from jsi::Value to the two new value wrappers
- Added conversion to jsi:Value from value wrappers.