Add global method that allows manipulation of states in Gesture Handler (#2519)
## Description

This PR adds a new `_setGestureState` method to the global object which will be used by the `react-native-gesture-handler`. In order to accomplish synchronous communication between modules I added shared interfaces between libraries (`GestureHandlerStateManager.java` and `RNGestureHandlerStateManager.h`). On startup RNGH module is tried to be loaded and then casted to the correct interface.
I did not update the documentation since this method is intended to be used internally.

## Changes

- Added global method `_setGestureState`
- Added loading of RNGH module and shared interfaces
- Added `Custom` gesture to the relevant Babel plugin list
- Changed `onBegan` to `onBegin` in the Babel plugin

## Checklist

- [ ] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Added TS types tests
- [ ] Added unit / integration tests
- [ ] Updated documentation
- [ ] Ensured that CI passes