Fix bezier & bezierFn mock in ReanimatedV2 (#3034)
## Description

After https://github.com/software-mansion/react-native-reanimated/pull/1879 was implemented (and also #2792), the mocks weren't updated. This PR fixes that.

The error is the same as in https://github.com/software-mansion/react-native-reanimated/issues/2766, but only shows up when running tests.

## Changes

The mocks now match the signature of the functions.

## Test code and steps to reproduce

There wasn't anywhere obvious to add a test for this - as far as I can see, we don't have any tests that use the mocks. I'm happy to add one though if you point me in the right direction!

To reproduce, it's as simple as a jest (with mocks set up) test that includes code like:

```typescript
const someEasing = Easing.bezierFn(0, 0, 0.15, 1);
```

## Checklist

- [x] Included code example that can be used to test this change
- [x] Updated TS types
- [ ] Added TS types tests
- [ ] Added unit / integration tests
- [ ] Updated documentation
- [ ] Ensured that CI passes