Load `RNGestureHandlerModule` lazily on iOS (#3166)
## Description

At the time, the `RNGestureHandlerModule` is loaded when the `createReanimatedModule` method is called. This causes `RCTBridge required dispatch_sync to load RNGestureHandlerModule.` warning to appear. This PR changes it, so the module is loaded only the first time the `setGestureState` function is called.

## Test code and steps to reproduce

Install `react-native-reanimated` and `react-native-gesture-handler` and check if the warning appears.