Update community page in the documentation (#3463)
## Description

At the moment the images in the community page in the docs are a bit large. For some reason (I haven't investigated) the class applied to the `img` tag gets overwritten, so I added a `div` wrapping the images and applied the class to that element. Also updated the current team and hidden the arrow next to the GitHub logo in the navbar.

## Changes

Wrapped the images with a div with the correct class applied.
Updated the current team.
Hidden the arrow next to the GitHub logo in the navbar.

<!--
Please describe things you've changed here, make a **high level** overview, if change is simple you can omit this section.

For example:

- Added `foo` method which add bouncing animation
- Updated `about.md` docs
- Added caching in CI builds

-->

## Test code and steps to reproduce

Visit https://docs.swmansion.com/react-native-reanimated/docs/next/community/ and change version

<!--
Please include code that can be used to test this change and short description how this example should work.
This snippet should be as minimal as possible and ready to be pasted into editor (don't exclude exports or remove "not important" parts of reproduction example)
-->

## Checklist

- [ ] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Added TS types tests
- [ ] Added unit / integration tests
- [ ] Updated documentation
- [ ] Ensured that CI passes