Restore d.ts file to npm package (#3496)
## Description

The PR [#3365](https://github.com/software-mansion/react-native-reanimated/pull/3365) was partially reverted. We use Bob to generate JS for the web, but we need to have the old d.ts file because of some TypeScript types problems in the d.ts file generated by Bob. Unfortunately, we forgot about restoring the old `react-native-reanimated.d.ts` to the package.

Fixes #3487