Fixed retain cycle of `REAModule` causing memory leak (#3555)
A retain cycle of `REAModule` caused by some lambdas in `NativeProxy`
that held a reference to it are removed.

A mechanism for easily detecting such memory leaks in the future is also
introduced in the form of the `SingleInstanceChecker` class. When added
as a class field, it counts the number of its instances and crashes the
app with a warning if more than one instance is detected.

There is no need to pass `reanimatedModule` to the lambda. If we only
pass `nodesManager` (which is `reanimatedModule.nodesManager` of course)
then the retain cycle is broken.

A method was also added to `REAModule`:

```objective-c
+ (BOOL)requiresMainQueueSetup
{
  return YES;
}
```

This is because the `init` method was not present before and after
adding it, it caused this warning:

`Module RCTImagePickerManager requires main queue setup since it
overrides init but doesn't implement 'requiresMainQueueSetup'. In a
future release React Native will default to initializing all native
modules on a background thread unless explicitly opted-out of.`

Since Reanimated was initialized on the main thread before, we want to
keep it that way.

- [ ] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Added TS types tests
- [ ] Added unit / integration tests
- [ ] Updated documentation
- [x] Ensured that CI passes

Co-authored-by: Tomek Zawadzki <tomasz.zawadzki@swmansion.com>