Unify assertion for no multiple instances of Reanimated with RNGH (#3493)
See:
https://github.com/software-mansion/react-native-gesture-handler/pull/2174

Fixed crashing on symlinks with filtering files before executing
operations on them, added property `disableMultipleInstancesCheck `
(possible to specify in `gradle.properties` of **rootProject**) which
allow application to disable the feature.

Note that this property is shared with `react-native-gesture-handler` -
so if someone sets `disableMultipleInstancesCheck=true` the assertion
won't run for neither `react-native-reanimated` nor
`react-native-gesture-handler`.

Works fine on `TestsExample` application in `react-native-screens`
repository.

- [ ] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Added TS types tests
- [ ] Added unit / integration tests
- [ ] Updated documentation
- [ ] Ensured that CI passes

Co-authored-by: Jakub Piasecki <jakub.piasecki@swmansion.com>