Fix iOS Warning: RCTBridge required dispatch_sync to load REAModule (#3698)
## Description

Fixes #3678. Reverts part of #3555 for Paper.

On Paper, the warning "RCTBridge required dispatch_sync to load
REAModule" was gone when I removed `requiresMainQueueSetup` method
introduced in #3678. However, the PR description explicitly states that
this method was added in order to eliminate the following warning:

> This is because the `init` method was not present before and after
adding it, it caused this warning:
> 
> `Module RCTImagePickerManager requires main queue setup since it
overrides init but doesn't implement 'requiresMainQueueSetup'. In a
future release React Native will default to initializing all native
modules on a background thread unless explicitly opted-out of.`

However, after the removal of `requiresMainQueueSetup` no warning
appears, so I assume that it is in fact not necessary to initialize
REAModule on main queue, thus it's safe to remove
`requiresMainQueueSetup`.

On Fabric, removing `requiresMainQueueSetup` it crashes the app on
reload, so I've left it as it is for now. We may address this
inconsistency in a separate PR.