Align Example app template with 0.70.5 (#3750)
This PR aligns Example app with RN 0.70.5 app template in order to simplify the process of adoption of RN 0.71.0 (#3745).

Even though Example never runs with Fabric enabled (it's Paper-only), I decided to leave new-arch Java and C++ files which will be deleted in 0.71.0-rc.0 for the sake of completeness.

Example app has a few dependencies, e.g. SVG, PagerView or MaskedView - as of this PR they are autolinked instead of being linked manually in `settings.gradle`.

I also removed some files specific for tvOS since we already have a separate app called TVOSExample.

- [x] remove `react-native-gradle-plugin` from Example/package.json
- [x] check Android / iOS
- [ ] check debug / release
- [x] check debugging worklets (Flipper needs to be launched before the app)
- [x] check if `MaskedView` component works