Move @types/invariant to devDependencies (#3788)
## Description

This PR moves @types/invariant to devDependecies as suggested by @hendryems in #3770.

Since there is only one occurrence of `invariant` in our codebase, first I thought we can replace this check with some simple JS code (if + throw new Error). However, invariant is also a dependency of react-native, so it makes no difference in terms of number of dependencies if we also rely on it.

## Changes

- Move @types/invariant to devDependencies

## Test code and steps to reproduce

Just see if everything works before release.