Fix Fabric crash on reload on iOS (#3789)
## Description

This PR fixes iOS crash on reload of FabricExample app. Most likely the regression was caused by #3332.

## Changes

First, I observed that the crash was gone after I commented out the following line:

https://github.com/software-mansion/react-native-reanimated/blob/1bec9a069371bc27a3319d8c0a7aa31b77a98de7/ios/REAModule.mm#L159

Next, I confirmed that `removeEventListener` was indeed called in `invalidate` here:

https://github.com/software-mansion/react-native-reanimated/blob/1bec9a069371bc27a3319d8c0a7aa31b77a98de7/ios/REAModule.mm#L70

However, this method was called on the ShadowQueue thread while `addEventListener` was called on the JS thread. As this may potentially cause a race condition when accessing internal `std::vector` of EventListeners, I decided to wrap the call with `runtimeExecutor` in order to schedule the call on JS thread, like here:

https://github.com/software-mansion/react-native-reanimated/blob/1bec9a069371bc27a3319d8c0a7aa31b77a98de7/ios/REAModule.mm#L140-L148

This resolved the issue. However, when I set a breakpoint on `removeEventListener`, it was never hit, so the method was actually never called, so I decided to completely remove `removeEventListener` call. After the changes, reloading app works fine, even while `Animated.ScrollView` is being scrolled.

## Test code and steps to reproduce

Launch FabricExample, open useScrollViewOffset example and reload the app a few times while scrolling.