Fix animated refs on Fabric (#3822)
## Summary

This PR fixes crash on Fabric when using synchronous measure/scrollTo
etc – this error was a regression introduced in #3722.

The issue was that did not transform shadow node wrapper ref into
shareable ref. Shareable ref was necessary for the UI side to recognize
and extract the shadow node wrapper that was needed for the sync calls
to execute.

## Test plan

Run FabricExample, use "Measure on UI" on the "Measure Example" screen
and "ScrollTo on UI" button on the "ScrollTo Example" screen.