Remove _setGloalConsole JSI method (#3853)
After #3838 we can now access the main global object via `global`
variable. This eliminates a need for _setGlobalConsole JSI method that
we'd install in the UI runtime only so that we can set the console
object as global on the UI runtime. In this PR we change the call to
_setGlobalConsole with just reassigning global.console.

Add some logs in worklets, run the app, see the logs appear on the
output, also that console.warn displays in the log box.

Co-authored-by: Tomek Zawadzki <tomasz.zawadzki@swmansion.com>