[Layout Animations][iOS] Fix registering exiting view ancestors (#3849)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->
#3824 introduces a bug with how we register ancestors of views with
`exiting` animations.
When a tree with `exiting` views is reattached to the native view
hierarchy, we don't register the ancestors of that tree as ancestors of
exiting subviews. This may cause issues with deleting these views early,
while there're still `exiting` animations running inside them.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->