Rewrite `isAnimated` function (#3949)
## Summary

This PR provides a new better implementation of `isAnimated` utility
function as the previous version was very poorly written. The old
implementation holds a queue of props to be checked and thus needs to
allocate and populate a new array on each call, whereas the new
implementation uses recursion instead &ndash; however, in most of the
real-world cases, it is never called recursively. Since `isAnimated` is
used very frequently throughout the codebase, previously it took ~7% of
total execution time. After the changes, it dropped down to ~3%.

Co-authored-by: @kmagiera

## Test plan

Just check if BokehExample works.