[Android] Fix wrong variable name in build.gradle (#4008)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->
This PR fixes an inexistent variable being used in a warning message in
`android/build.gradle`. If condition for this warning was matched, the
Android build would fail.

I'm supposing that the use of `rnMinorVersion` variable is the correct
replacement in this case.

EDIT:
Found an issue from last month reporting this problem, so this PR fixes
#3861