Fix Fabric crash in `NativeReanimatedModule::removeShadowNodeFromRegistry` (#4020)
## Summary

This PR fixes a Fabric crash in
`NativeReanimatedModule::removeShadowNodeFromRegistry`. I don't know the
exact root cause but I know how to fix it.

First, I added missing `makeShareableShadowNodeWrapper` call. However,
since only the view tag is required on the native side, there's no need
to pass the whole ShadowNodeWrapper, so I simply grab the view tag from
`this._viewTag` and pass it to the cleanup worklet via closure.

## Steps to reproduce the issue

1. Launch FabricExample on iOS
2. Open "forwardRef & useImperativeHandle" example
3. Increase animation duration to 3000 ms
4. Randomly click both buttons
5. Go to previous screen
6. Crash


https://user-images.githubusercontent.com/20516055/216365412-edb406ad-6947-4cdf-94aa-5b82bd079fab.mp4

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->