[Shared Element Transition] Improvements after review v1 (#4014)
## Summary

This PR contains improvements after the latest reviews form PR with
[Shared Element
Transition](https://github.com/software-mansion/react-native-reanimated/pull/3827)

Changes:
- Naming improvements
- Get rid of the key-value observer - In the previous implementation I
used KVO to change the context of execution. That was because swizzled
method changes the context of execution, so the self in the swizzled
method is `RNSScreen` instead of `REASharedTransitionManager`. That's
why I used KVO to change the context of execution back to
`REASharedTransitionManager`
- I used directly `[RNSScreen class]` and `[RNSScreenView class]` in the
`swizzleScreensMethods` method, instead of getting it from the instance
of screen (`[screen class]`)
- Refactor Snapshot on iOS