Layout animations cleanup (#4090)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
Clean up the layout animation rewrite from #3332

- use Enum types for layout animation types ('entering', 'exiting' and
'layout') in both JS and native parts of the implementation
- try caching `AnimationsManager` in `LayoutAnimations.java` instead of
calling `getNativeModule` (Android implementation)

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->