Update README.md (#4127)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

In honour of V3, i just updated the ReadMe to say Reanimated 3 :)

## Test plan
No specific test plan, but i did see some broken links in the docs, so i
might do other PRs for those after this