Rewrite plugin to TypeScript (#4096)
## Summary

Rewriting plugin to typescript. Plugin will compile when doing either
`yarn` or `yarn plugin`, hence it will be available in npm tarball.

I think the best way to review the changes is to compare the compiled
'index.js' file with the old one (having in mind compilation options in
'tsconfig.json'), then remove it in the follow-up commit.

## Test plan

Generate bundles for `Example` and `FabricApp` for `main`.
Generate bundles for current head of this branch with freshly compiled
`index.js`. Compare the bundles. There should be no differences as long
as they were built in the same directory.

---------

Co-authored-by: Tomek Zawadzki <tomasz.zawadzki@swmansion.com>