Fix issue causing ANR when performing sync layout updates (#4239)
## Summary

This PR addresses the problem from #3879 on iOS where application would
fall into a deadlock when performing synchronous layout updates. The
reason for the issue was that sync layout updates requires UI thread to
lock and wait for layout thread to perform some tasks. On the other
hand, layout thread sometimes schedules UI tasks to be performed
synchronously (with `dispatch_sync` mechanism).

The solution this PR implements is to add a timeout on waiting and
fallback to updating layout props asynchronously. This feels like a
right approach as the cases when dispatch_sync is used are relatively
rare and the consequences of updating layout asynchronously aren't
serious (there will be a one frame delay for that one update).

This solution is based on #3082 but on top of just adding the timeout we
also need to make sure that the executed operations are properly
mounted. The problem with #3082 was that when the UI thread times out we
still use `runSyncUIUpdatesWithObserver` call on the layout thread which
in turn results in the mounting block being stored and expected to be
run on the locked UI thread afterwards. If the thread times out we may
hit an issue when the mounting block is set but never executed which
would have serious implications.


## Test plan
Tested this with repro scenario provided in #3946 that relies on
`@react-native-community/datetimepicker` library