Fix monorepo workflow (#4273)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

There seems to be a problem with running `npx react-native` from inside
a directory that contains package.json file (see
https://github.com/facebook/react-native/issues/34441#issuecomment-1275186909),
however I wasn't able to reproduce it locally.

This PR swaps the order of steps, so that `npx react-native` is run
before package.json is created.

## Test plan

Just check if CI is green.