Fix "Maximum call stack size exceeded" error (#4367)
## Summary

This PR addresses a number of issues reported related to exceeding stack
when converting objects to shareables (e.g. in #4177).

This issue got introduced in Reanimated 3 after #4060 where we got rid
of filtering only attributes accessed in worklets to be captured by
worklets.

This PR implements a solution that filters only plain objects before
converting things to shareables. In all of the instances of this issue
we encountered, the crash was due to the fact we were leaking some react
internals (fiber nodes) into the captured variables (e.g. via React's
refs or otherwise). These objects are not of a plain object type and
hence we wouldn't be able to support transferring these on the UI
runtime anyways. With this change we are skipping these type of objects
and replace them with a "inaccessible object" which is implemented as a
proxy that throws on any attempt of accessing its attributes – this way
users will get an error if they do use or access such objects.

In addition we add a cyclic object detection mechanism. Currently cyclic
objects cannot be converted to sharables – they weren't supported in
Reanimated 2 either. We add it as a part of this change in case the
object prototype check wouldn't be sufficient in some cases. In a
scenario when cyclic object is captured by a worklet we will detect that
and throw a more descriptive error.


## Test plan

The below code illustrates a scenario when we are capturing react
internals via ref:
```js
function Mleko({ props }: any) {
  const style = useAnimatedStyle(() => {
    console.log(props)
    return {}
  })
  return (
    <View ref={props.ref} />
  );
}

const ref = useRef();
<Mleko props={{ ref }} />
```

Before this change the above coud would throw stack exceeded error. Now
this works ok.

In addition we tested cyclic detection by capturing the following
object:
```js
const one = {};
const two = {one};
one.two = two;
```

The above object would throw stack exceeded error. now it throws an
error that says that user is capturing cyclic object.


Lastly, we tested accessing non plain objects:
```js
class Something() {
  constructor() { this.ggg = 7 }
}
const something = new Something();

useAnimatedStyle(() => {
  console.log(something.ggg);
  return {};
});
```

---------

Co-authored-by: Tomek Zawadzki <tomasz.zawadzki@swmansion.com>