Refactor keyboard observer on iOS (#4360)
## Summary

This PR refactors the logic in REAKeyboardEventObserver to address some
number of different issues:
1) no change notification when keyboard size changes due to change in
language or showing suggestions bar
2) issue with keyboard height notification when opening modal (see
#4339)
3) issue with non-zero height after keyboard hides completely.

Here is the list of the main thing that this refactor changes:
1) We no longer use willShow/willHide/didShow/didHide notifications for
keyboard and only use willChangeFrame which is the most reliable one and
triggers also when the keyboard updates
2) We no longer use a technique where the keyboard size is determined
based on the frame of the keyboard view. This was problematic as that
frame would sometimes reflect a one-frame delayed size which resulted in
the animation not coming down to 0 or not updating at all when
immediately hidden or appear (by a modal). Instead, we use UIView
animation to animate a complementary view that is placed outside of the
screen visible area. We use UIView animation to animate height of that
invisible view and then use display link to access presentation layer in
order to read up-to-date height.

There are few gatchas related to the new technique used:
1) Because we only look at frame change notification we need to properly
determine the state of the keyboard. We do it based on the target
height. If it is 0 then we consider the state to be either closing or
closed – we select the final state based on whether there is animation
running (closed state when no animation, or closing when we are
animating). This prevents us from changing state to closing when
keyboard is only becoming shorter (e.g. when text suggestion bar is
closed). We handle opening/open state in similar way.
2) Due to the UIView animation technique we need to check
presentationLayer.animationKeys to determine if the layer runs the
animation. If it doesn't then we extract height from the view frame and
not from the presentationLayer frame, as the presentationLayer may not
have an up-to-date information (it will be delayed by one frame).

## Test plan

Use example from #4339 and animated keyboard screen from Example app.
Open/close text suggestion bar to see how the app reacts when keyboard
is changing height and not just opening/closing.