Set the opposite sensor values on android to be consistent with iOS (#4368)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

The sensors on Android and iOS return opposite values, which cause
inconsistencies. To fix this, I modified the Android sensor listener to
return opposite values.
I didn't modify rotation values, because these are the same.

## Test plan

- Ran each sensor and compared the values returned by both platforms
- Tested an example application to confirm that the behavior of the
sensors is consistent across both platforms.