Remove commonjs generation (#4468)
## Prelude

We have waited for it a long time. 

Here it is.

A perfect summary of weeks time of our work and research - two line fix.

## Summary

Ever since we've adopted CommonJS as target for our built JavaScript
files we made it very difficult to our users to properly use Jest with
Reanimated. This is due to fact that Reanimated Babel plugin is prone to
break hoisting and this becomes critical in CommonJS. This issue had
never surfaced before because we were building only for ES modules which
can handle the way our plugin works. Therefore, after many talks we've
decided to drop support for CommonJS and revert to ES modules only.

Refer to #4449 for _slightly_ more information.

This is likely to fix (I want to wait for confirmation):
- #4347,
- #4197,
- #4185,
- #4175,
- #4136,
- #4001,
- #3971.

## Test plan

- I tested it on the [original
repro](https://github.com/kylebake/ReanimatedTestBug) that began this
voyage. I recommend using [my
fork](https://github.com/tjzel/Reanimated-jest-commonjs-issue) since
that repro had some errors coming from used versions of React-Native and
Jest.
__Make sure you use Node 18.15 for it.__
- I also tested it on my university
[app](https://github.com/MeowdoptMe/MeowdoptMeApp) that actively uses
Jest and Reanimated and using tarball from this PR fixed Jest issues.