Update `useAnimatedSensor` types (#3848)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR is part of an ongoing effort to get rid of `.d.ts` file and use
type definitions generated based on the source files.

In this case, I've changed `AnimatedSensor` to be generic and provide
different sensors depending on what it's parametrized with. This way
it's easy to overload `useAnimatedSensor` and setup the correct typings
depending on what sensor is being created.

---------

Co-authored-by: Tomasz Żelawski <tomasz.zelawski@swmansion.com>