Update types in layout animation builders (#3833)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR is part of an ongoing effort to get rid of `.d.ts` file and use
type definitions generated based on the source files.

The focus of this PR is on the code responsible for building layout
animations. Instead of hard-coded return types in builders, it uses the
implicit `this` argument to infer the type that should be returned.