Update types of some animations (#3850)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR is part of an ongoing effort to get rid of `.d.ts` file and use
type definitions generated based on the source files.

The main focus of this PR is to make `withDelay` and `withRepeat`
generic functions so that they can accept any animation as an argument
instead of only their own type.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->