Add scrollViewOffset prop to AnimatedScrollView (#4414)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR moves logic from `useScrollViewOffset` hook to
`Animated.ScrollView` prop.

The API of `useScrollViewOffset` is a bit cumbersome, because we need to
pass the same `ref` to both `ScrollView` and hook.

Now we just need to pass variable with type `ScrollValue<number>` to
`Animated.ScrollView` and when Scroll View changes its position, our
value is updated

## Test plan

- Example app