Memoize ReanimatedModule in `sendEvent` (#4525)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR adds `static __weak REAModule *reaModule;` variable which stores
the result of `[[self bridge] moduleForName:@"ReanimatedModule"]` so we
can call it only once during initialization instead of on each event.

When `REAModule` is deallocated, all `__weak` pointers are automatically
replaced with `nil` so it doesn't leak any resources as well as works
fine after a reload.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->