Set default scrollEventThrottle (#4589)
## Summary

The inspiration for this changes are two issues:
-
https://github.com/software-mansion/react-native-reanimated/issues/4585#issuecomment-1595921845
-
https://github.com/software-mansion/react-native-reanimated/pull/4416/commits/4137d1a7d4ddba180c54582c2e1036a285c0ba3e

Users expect continuous scroll events. But the default value of
[scrollEventThrottle](https://reactnative.dev/docs/scrollview#scrolleventthrottle-ios)
is 0, which results in the scroll event being sent only once each time
the view is scrolled. Which is easy to forget about it and in effect, it
seems like a Reanimated issue. I know that this is a change of default
ScrollView behavior but I believe that it would be less confusing for
the users.