Use composition over inheritance for `RuntimeManager` (#4608)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Currently, `NativeReanimatedModule` extends `RuntimeManager` which is
quite inconvenient as we would like to have multiple instances of
`RuntimeManager` in the future (and also merge this class with
`JSRuntimeHelper`).

This PR converts inheritance into composition, i.e. adds
`std::shared_ptr<RuntimeManager> runtimeManager_;` field to
`NativeReanimatedModule` class as well as updates all occurrences where
we use it.

## Test plan

Check if Example and FabricExample compiles successfully on Android and
iOS.