Unify naming of `NativeReanimatedModule` and `REAModule` instances (#4612)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR aims to rename `module` variable which holds an instance of
`NativeReanimatedModule` because `module` is a keyword since C++20. From
now on, we will use the following naming convention:

* `std::shared_ptr<NativeReanimatedModule> nativeReanimatedModule`
* `std::weak_ptr<NativeReanimatedModule> weakNativeReanimatedModule`
* `REAModule* reaModule` (iOS)

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->