Fix type tests (#4667)
## Summary

Type tests have been neglected for quite a while but they are
nevertheless useful and now it's the time to give them some more love.

- I have removed `react-native-reanimated` import in favour of relative
import. Once in the past I debated about this with @tomekzaw that the
former import was also checking if importing with module name was
correct, but most of the time my TypeScript was having great troubles
with resolving it and typing things inside the tests as `any` which
ruins their purpose. Therefore I think it's much more useful that way.
- I have added api check to our types CI
- I have refactored `AnimateProps` type to get rid of included type
recursion. Also I added a type alias `AnimatedProps` for it since this
should be a proper name for it.

## Test plan

Check that type tests have less errors than before. They still have
them, it's to be improved in the future.

---------

Co-authored-by: Alex Cynk <ola.zn123@wp.pl>