Fix inline styles for Animated Components (#4671)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Fixes #4584 

Currently we incorrectly provide an error if a user tries to use
`sharedValue` or `derivedValue` as a style property. This PR fixes it.

## Test plan

Added relevant tests.