Fix TypeScript issue with useAnimatedProps (#4653)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

During the migration from manual types, the proper typing of
"useAnimatedProps" has proven to be quite challenging. As a result, I
had to resort to casting it to its previous type, but unfortunately, I
missed some places.

## Test plan

In our new docs the given
[example](https://reanimated-beta-docs.swmansion.com/docs/core/useAnimatedProps#adapters-)
doesn't work.
After applying these changes it's fixed.