Update app icon and splash screen in Example (#4686)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

While working with app icon and splash screen for FabricExample app (see
#4676), it turned out that the approach used in Example is outdated.
This PR fixed this and makes both approaches similar.

iOS:
From now on there is only one icon image file. There is also a new file
`Launch Screen.storyboard` that replaced `LaunchScreen.xib`. The
appearance of the icon and the launch screen remained the same.

 


https://github.com/software-mansion/react-native-reanimated/assets/56112901/c64577be-a1e9-493a-b143-5911c6744c4e



Android:
`res/` directory has been restructured to make it similar with
FabricExample. The splash screen was implemented with

[androidx.core:core-splashscreen](https://developer.android.com/jetpack/androidx/releases/core)
thus the screen is present in Android 11 and lower. The appearance for
newer Androids didn't change.



https://github.com/software-mansion/react-native-reanimated/assets/56112901/a1f99cf2-9f64-4cc0-9cbd-ad69bbb3c2fb



https://github.com/software-mansion/react-native-reanimated/assets/56112901/5808dd76-be93-42e1-b6fe-4d928e9211e5



<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->