Swizzle RCTUIManager (#4619)
## Summary

Recently, we began rewriting the iOS initialization process. Tomek moved
the injection of JSI bindings to `installTurboModule` (check out
https://github.com/software-mansion/react-native-reanimated/pull/4523)
and then used swizzling to eliminate the `REAEventDispatcher` (see
https://github.com/software-mansion/react-native-reanimated/pull/4576).
More context here:
https://github.com/expo/expo/pull/23057#discussion_r1240070245

This PR removed UIManager overriding by swizzling the `_manageChildren`
and `uiBlockWithLayoutUpdateForRootView` methods. These changes allowed
us to get rid of `RCTAppDelegate+Reanimated`, `UIResponder+Reanimated`,
and `REAInitializer`.

For now, we're gonna leave `REAInitializer` as deprecated for a few more
releases to maintain backward compatibility. But just a heads up, we'll
be removing it in the future.

## Test plan
- I tested building on different react native versions by CI
https://github.com/piaskowyk/reanimated-rn-version-tester/actions/runs/5375166823
- Tested Layout Animations on Example app form main branch.