Add useReducedMotion hook (#4687)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

Animations can cause some people to feel unwell. In order to help
developers create apps with those people in mind this PR adds a
`useReducedMotion` hook that returns a boolean value informing whether
the "Reduced Motion" setting was enabled when the app started.

https://github.com/software-mansion/react-native-reanimated/assets/56109050/6ffa6e06-2a29-499b-8119-409c943050dd

Test the behavior of ReducedMotionExample on android, iOS and Web.