Add duration support to springify (#4710)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->
Since we have added support for duration to our spring animations we
should do the same thing with `.springify()` modifier.

Now you can use the following syntax:
```
        entering={SlideInLeft.delay(500).springify(5000).mass(2)}
        entering={SlideInLeft.delay(500).springify(5000)}
        entering={SlideInLeft.delay(500).springify(5000).dampingRatio(1)}
```

Although we use chaining ⛓️ instead of config to pass other spring props
(like mass or stiffness) doing the same with duration would be unnatural
as we specify duration of delay as a function argument. Also it could be
confusing, since user could wonder if duration applies to spring or
delay.

## Test plan
Tested on our [LA] Carousel exampel
<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->



https://github.com/software-mansion/react-native-reanimated/assets/56199675/fdd71b63-393a-418b-bd86-cb6ab5c72c85

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: Aleksandra Cynk <aleksandracynk@swmansion.com>
Co-authored-by: Bartłomiej Błoniarz <56109050+bartlomiejbloniarz@users.noreply.github.com>
Co-authored-by: Jakub Fedak <56112901+jfedak@users.noreply.github.com>
Co-authored-by: Krzysztof Magiera <krzys@swmansion.com>
Co-authored-by: Tomasz Żelawski <40713406+tjzel@users.noreply.github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>