Rename `animatedRuntime` to `uiRuntime` (#4731)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR renames `animatedRuntime` to `uiRuntime` and `wrt` with
`weakUiRuntime` as well as replaces old dirty `*wrt.lock()` with a
proper check.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->