Mock version and source map in plugin snapshot tests (#4733)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

I noticed that `version` and `sourceMap` properties are not included in
plugin snapshots. This PR:

* renames environmental variables used for testing purposes
* unifies `compact` setting across app and tests and updates snapshots
accordingly
* restores `version` property in snapshots but uses mock version "x.y.z"
instead of real version
* moves `version` property to `__initData` to prevent allocating
multiple strings (see #3930 for explanation)
* restores `sourceMap` property in `__initData` but uses mock source map
for all snapshot tests
* adds one unit test that checks real `sourceMap` property
* adds double negation to fix return type of `isRelease` (previously it
returned `boolean | string | ""`)


## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->