Detected debug version of the worklet in release bundle (#4732)
## Summary

Fixes:
https://github.com/software-mansion/react-native-reanimated/issues/4690

It may happen when one of the user dependencies provided transpiled code
with debug version of the Reanimated plugin. In effect, we have a debug
version of the work in your release bundle. This might lead to
unexpected issues or errors. 😢

I created issue with description how to fix this issue:
https://github.com/software-mansion/react-native-reanimated/issues/4737

---------

Co-authored-by: Tomek Zawadzki <tomasz.zawadzki@swmansion.com>