CI for iOS build with USE_FRAMEWORKS=static (#4735)
## Summary

In this follow-up to the PR
https://github.com/software-mansion/react-native-reanimated/pull/4718.
This PR contains a CI test to check the Reanimated build with a static
framework configuration. The initial problem was reported in the
following link:
https://github.com/software-mansion/react-native-reanimated/issues/4701.
For this CI, two apps were created: Paper and Fabric architecture. The
CI setup involves using the command `USE_FRAMEWORKS=static NO_FLIPPER=1
&& pod install` to configure the app with a static framework. Since
there is no need to run this CI for every PR, it will be performed only
as a nightly test.