Rename confusing variable (#4779)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Currently, in our code we have
```ts
const IS_WEB = shouldBeUseWeb();
```
which is confusing because there's also a function called `isWeb()`.

This PR replaces all occurrences of `!IS_WEB` with `IS_NATIVE` like we
already do in a few places.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->