Add missing underscore to `_closure` and correct transform matrices test location (#4803)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

To be consistent with worklet properties we turn `_closure` into
`__closure`. Also I noticed that tests for transform matrices are not in
`__tests__` in root directory so to not get them dispersed I moved them
there.

## Test plan

Try out examples