Add ts-prune (#4772)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

* Ts-prune allows us to detect unused typescript code. 
* Advantages:
* We can remove exports of all the functions and interfaces exported for
no reason. This makes importing with cmd+enter shortcut easier, since
you are less likely to get wrong suggestions from vscode.
* We can see all the “forgotten” areas of code, and think about reasons
to have them.
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

> **Note**
>I've removed the file `MapperRegistry.ts` which seems unnecessary since
this PR:
[#3722](https://github.com/software-mansion/react-native-reanimated/pull/3722)


## Test plan

Just CI
<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Aleksandra Cynk <aleksandracynk@swmansion.com>