Bump React Native to 0.72.3 (#4867)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR bumps react-native to 0.72.3 in Example and FabricExample apps.

Turns out `Device.flow.js` from `metro-inspector-proxy` is unused so I
only updated the patch for `Device.js` and removed diff for
`Device.flow.js`.

## Test plan

Check if CI is green.