Update globals in plugin (#4868)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR adds the names of missing standard built-in objects (e.g.
`BigInt` or `WeakSet`) in Reanimated Babel plugin as well as removes
redundant ones (e.g. `LayoutAnimationRepository` or `__ErrorUtils`).

Based on
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects.

## Test plan

Verify each name one-by-one.