Remove Reanimated 2 logic from build.gradle (#4846)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR removes legacy stuff from build.gradle which is no longer needed
for Reanimated 3.

Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>