Unify AnimateStyle types (#4855)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Currently we have two `AnimatedStyle` types - one named `AnimateStyle`,
generic and present in previous .d.ts and other internal one, named
`AnimatedStyle` that is just one big `any`. This PR retains only the
first one and provides a proper alias for it.

`any` use as a generic argument in unified AnimatedStyle stems from its
previous use and will be changed and fixed in the future.

## Test plan

CI