Add `yarn find-unused-code:js` to pre-commit (#4875)
info Visit https://yarnpkg.com/en/docs/cli/run for documentation about
this command. to pre-commit

<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
Add `yarn find-unused-code:js` to pre-commit.
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan
Commit changes and see that nothing explodes :boom: 

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Co-authored-by: Aleksandra Cynk <aleksandra.cynk@swmansion.com>