Fix web bundling warning (#4892)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR removes the following warning that appears when bundling
WebExample:


![warning](https://github.com/software-mansion/react-native-reanimated/assets/20516055/c0cf5200-09bb-40ba-ba3d-a207b9995050)

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->