Move empty example at top and add `userSelect: none;` (#4895)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

This PR does two things:

1. Moves empty example at top of the main list, so it is easier to
access while we test new things
2. Adds `userSelect: none;` to DefaultLayoutAnimations example, because
sometimes it blocks exiting animation

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Tested on example app

Co-authored-by: Michał Bert <michal.bert@swmansion.com>