Removed unnecessary copies of UIManager and ContextContainer (#4812)
## Summary

I noticed that in our `ReanimatedCommitHook` on each commit we copy
`shared_ptr` of `UIManager` and `ContextContainer`. I've gone through
the code and such copies are not needed as object using them are
short-lived without option to deallocation of needed resources.

Those copies can be pretty time consuming due to locks in `shared_ptr`
implementation.

## Test plan

Check performance on Android and iOS