Lock `weakNativeReanimatedModule` using conditionals (#4871)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR unifies the way how we lock `weakNativeReanimatedModule` inside
platform-specific lambdas. Previously, sometimes we would first call
`.lock()` and then check its result with a conditional. However, the
shared pointer is available in the scope even if locking was not
successful. I would like to switch to syntax `if (ptr = weakPtr.lock())`
which prevents this inconvenience.

Additionally, I added `jsi::Runtime &` as the first argument to
`progressLayoutAnimation` function which can be bound automatically so
that we don't need to care about locking weak pointer to the runtime in
this particular lambda.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->