_log for complex types (#4857)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR expands the logic of _log function to print more complex types
in JSON-like style. It adds support for the types listed below:

- JSI types
- JS Objects and Collections

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

You can use and example app named `Log test`

---------

Co-authored-by: Michał Mąka <62388446+michalmaka@users.noreply.github.com>
Co-authored-by: Tomek Zawadzki <tomekzawadzki98@gmail.com>