Fix `findDOMNode` deprecated on web inside `StrictMode` (#4866)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

When `StrictMode` is enabled, web version of reanimated shows error that
`findDOMNode` is deprecated. This error comes from usage of
`findNodeHandle` function. This PR removes usages of `findNodeHandle` on
web version, so that error no longer appears.

This PR also introduces a lot of type casting. On web version
`_component` and `viewTag` are `HTMLElements` (i.e. divs that we
animate), hence I had to change types of these properties to also
include `HTMLElement`. It resulted in many type casts. Casts to `number`
are used mainly because code that requires `number` is not called on
web.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Tested on our example app.