Add `reduceMotion` method to animation builders (#4884)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
This PR extends the `BaseAnimationBuilder`, `Keyframe` and
`SharedTransition` with a `reduceMotion(config)` method. Possible values
of config are `ReduceMotion.System`, `ReduceMotion.Always` and
`ReduceMotion.Never`. It defaults to `system`.

In `BaseAnimationBuilder` and `Keyframe` the config is passed down to
animation functions.

Exiting animations and shared transitions are completely removed when
the config is set to `always` or it is set to `system` and the `Reduce
Motion` accessibility setting is enabled.
### LA example

https://github.com/software-mansion/react-native-reanimated/assets/56109050/5f618e78-83a4-4c47-ae3a-405be1cf24dc

### SET example

https://github.com/software-mansion/react-native-reanimated/assets/56109050/bfec2de3-cd6d-49cd-8efb-5ebd3d9508fa

(The disappearing text and other problems in this example are linked to
some issues with custom animations in SETs)

## Test plan

Test the behavior of the example app with the accessibility setting on
and off.