Fix text shadow not animating (#4967)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Since react-native-web 19 `textShadow` props are not animating properly.
This PR aims to fix that problem.

Fixes #4903 

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Tested on repo from [this
issue](https://github.com/software-mansion/react-native-reanimated/issues/4903)
on different react-native-web versions