Fix `useAnimatedProps` on web (#4977)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR changes behavior of `useAnimatedProps` so that props get updated
properly on web.

Fixes #4909

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Tested on [example from
issue](https://github.com/apatozi/issue-reanimated-useanimatedprops-web).