Fix scrollTo crash (#4984)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR adds check whether `animatedRef` has been set or not. If ref is
not defined, calling `element.scrollTo` results in crash, because
`element` is set to `-1`.

Fixes #4424 

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Tested on snack from
[issue](https://github.com/software-mansion/react-native-reanimated/issues/4424)