Fix `AnimatedStyle` for good (#4923)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

After many attempts to hotfix `AnimatedStyle` and always finding
something faulty about it I decided to rewrite it to make it closer to
RN types.

Key changes here are in `helperTypes` file - I removed manual types for
`AnimatedStyle` for a more generic approach. I included two utility
types: `MaybeSharedValue` and `MaybeSharedValueRecursive` which take a
generic argument and if it's an object - it makes its properties
possibly SharedValues, if it's an array - the same for its items, and if
it's a primitive - Shared Value of a primitive, if it's supported as
declared in `AnimatableValue` type.

With this change, `AnimatedStyle` is just recursively processed `Style`
of a given component and no more logic is required.

Consequently I fixed typing of `useAnimatedStyle`, removed type casts
from there and used TS function overloading to differentiate between
user usage and developer usage (look at the code).



## Test plan

I added 500 lines of type tests, many of them fail without this PR.

## Note

During work on this PR I noticed that not all viable usages of inline
Animated Styles are supported - I will work on it in the future.