Fix crash on reload in `ShareableRemoteFunction` (#5015)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR fixes a crash that occurs during a reload of the app.
`ShareableRemoteFunction` holds a reference to a remote function in
`function_` field. When `ShareableRemoteFunction` is deallocated,
`function_` also gets deallocated. If the runtime has already been
terminated, the assert in `~jsi::Pointer` will fail.

![Zrzut ekranu 2023-08-31 o 10 48
42](https://github.com/software-mansion/react-native-reanimated/assets/20516055/3685ce3c-639d-42c5-84da-a4f238d2cf19)

## Test plan

Tested on VisionCameraExample.