Expose `makeShareableCloneRecursive` (#5018)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR adds `makeShareableCloneRecursive` to the public API.

Example use case:

```ts
const runtime = createWorkletRuntime('foo');
const worklet = () => {
  'worklet';
  console.log(42);
};
someJSIBindingFromThirdPartyLibrary(runtime, makeShareableCloneRecursive(worklet));
```

We need to call `makeShareableCloneRecursive` here so that the argument
is an instance of `std::shared_ptr<ShareableWorklet>` and not just a
`jsi::Function`. Also, we need the whole worklet closure to be
shareable.

An alternative would be to expose a new function `makeShareableWorklet`.

## Test plan

See typetests.