Support Layout Animations on macOS (#4985)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

Previously, Layout Animations were disabled on macOS platform (see
#4912). This PR adds this functionality.

In order to make Layout Animations work on macOS, several changes had to
be made:

- add `center` property to RCTUIView - our initial idea was to just add
this property to REAUIView, but this approach was not sufficient;
- change the way of getting `absolutePosition` - according to Apple
documentation (see [Appkit
docs](https://developer.apple.com/documentation/appkit/nsview/1483406-convertpoint?language=objc)
and [UIKit
docs](https://developer.apple.com/documentation/uikit/uiview/1622442-convertpoint?language=objc)),
there is no need to pass specific view to `convertPoint` method if we
want to get window based coordinates. This is a correct approach on
macOS, since we can't pass NSWindow to this method;
- use different method to get window's size on macOS - instead of
`bounds.size` we used `frame.size`;
- set `header = nil` on macOS - native stack is not available on macOS,
thus we are sure the first view is not a header

## Test plan

Tested on macOS (working fine) and iOS (not breaking anything)

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->