Add fallback to setImmediate for Expo Router static rendering (#4665)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
- https://github.com/expo/router/issues/718
- [Twitter
thread](https://twitter.com/bhyoo99/status/1675302401677615104?s=20)

In Node.js environments such as when static rendering with Expo Router,
requestAnimationFrame is not defined. This results in an error when
using libraries that depend on react-native-reanimated like Drawer or
when using react-native-reanimated directly in Expo Router with static
rendering, causing a hydration mismatch due to rendering errors.

In this PR, I added a fallback to setImmediate if requestAnimationFrame
is not defined in `globalThis`. I did not simply replace
requestAnimationFrame with setImmediate because [`setImmediate` is
slower than
requestAnimationFrame](https://github.com/software-mansion/react-native-reanimated/pull/1521)
in CSR environments. This issue is not applicable to static rendering
environments however, so setImmediate is fine there.

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan
- See [this
issue](https://github.com/expo/router/issues/718#issue-1784125278)'s
Minimal reproducible example
<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Tomasz Żelawski <tomasz.zelawski@swmansion.com>
Co-authored-by: Tomek Zawadzki <tomekzawadzki98@gmail.com>