Improve entry exit animation types (#4998)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

1. Use typescript regex types to generate object with `current` and
`target` prefixes based on a single source of truth
- This may be a good change, since we consider supporting more animation
properties.
3. Extract common properties from similar types to reduce boiler-plate
code, use inheritance instead of copy-pasting them
4. Get rid of `[key: string]: number | number[];` 🥳 
5. Remove lots of type casts in `SharedTransition.ts` 🥳 


<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan
`yarn type:check` & `yarn type:check-api`

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Aleksandra Cynk <aleksandra.cynk@swmansion.com>