Fix `measure` crash on web (#5060)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

While going through our example app I stumbled upon crash in "Article
progress" example. Turns out that sometimes we try to call
`getBoundingClientRect` on element that has not been rendered yet. This
PR fixes this problem.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Tested on "Article progress" example.