Use strict everywhere (#5055)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

I'm not sure if we need this pull request, but it only took me 
```bash
#!/bin/bash
for file in $(find . -type f \( -name "*.ts" -o -name "*.d.ts" \))
do
  sed -i '' -e $'1i\\\n'"'use strict';"$'\n' "$file"
done
```
to make it.

It would help us in the past when @bartlomiejbloniarz ran into issues
when `Object.freeze` froze his object and he couldn't modify it - and
without `use strict` this is a silent fail.

## Test plan

Many clicks in Example app.