Revert #5051 (#5084)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Revert change
https://github.com/software-mansion/react-native-reanimated/pull/5051

We have decided to revert [this
change](https://github.com/software-mansion/react-native-reanimated/pull/5051)
since using such a fixes instead of understanding the reason of the
problem may lead to future bugs.

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->
 
Try to reproduce the initial issue

Co-authored-by: Aleksandra Cynk <aleksandra.cynk@swmansion.com>