[SET] Fix transition restart (#4947)
## Summary

This PR fixes the issue with shared element transition restart.
Previously, there was a chance of the app crashing or a needed component
disappearing. The purpose of this PR is to prevent app crashes and
ensure a proper end state for the transition.

## Test plan

Run example `TransitionRestart.tsx` example - available in this PR