Add example for Next.js (#5108)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR adds an example with Next.js and a CI to test build fails. It is
intended to prevent using api that is only available in a browser (w.g.
window) without an appropriate check. This ensures that reanimated can
be used with SSG.

## Test plan

Build the example with `yarn build` and start with `yarn start`. You can
also start it in dev mode using `yarn dev` and check if it responds
changes in reanimated core.