Add e2e tests in NextExample (#5124)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR adds e2e tests for web. Tests are implemented with cypress.

## Test plan
Check if the test passes, then change the target width in `test.js` and
check that the test doesn't pass. Execute `yarn build` to build
`NextExample` and then run `yarn e2e:headless` to start tests.