Prevent top-level calls to `requestAnimationFrame` (#5136)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Currently there are instances where `scheduleOnUiImmediately` is called
in the top level to create global structures which are not needed on
web. This results in a call to `requestAnimationFrame`, which is
unavailable during static website generation. We already have a mock for
`raf` which prevents the error, but in this PR I changed the logic so
the unnecessary calls to `raf` don't happen on web.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->