Add config component for layout animations (#5045)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR adds `<LayoutAnimationConfig>` - a wrapper component, which can
be used to
- disable entering animations on the first render of the wrapper
- disable exiting animations when the wrapper is removed

It also adds a `skipEnteringExitingAnimations` prop to
`Animated.FlatList` - when used entering and exiting animations of items
will not start when the list appears/disappears.


https://github.com/software-mansion/react-native-reanimated/assets/56109050/6b565e6e-1df5-4174-b0d4-1500fb3e5e88

## Test plan

Check the behavior of `FlatList skip entering & exiting` and `Nested
LayoutAnimationConfig` examples.