Get rid of native=true from ReanimatedModule (#5134)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan
Let's notce that `NativeReanimatedModule.native` is equal to
`!shouldBeUseWeb()`. We can simplify our code a little substituting it.

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Aleksandra Cynk <aleksandra.cynk@swmansion.com>