Inject `_updateDataSynchronously` to worklet runtime (#5172)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Previously, we would inject `_updateDataSynchronously` only to the UI
runtime so it was not available in a regular worklet runtime. However,
VisionCamera needs to use synchronous shared values in its context so I
moved `_updateDataSynchronously` from `UIRuntimeDecorator` to
`WorkletRuntimeDecorator`.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

```ts
const mutable = makeMutable(false);
const runtime = createWorkletRuntime('foo', () => {
  'worklet';
  mutable.value = true;
});
```