Fix JSISerializer for empty jsi::HostObject (#5183)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR fixes a crash when passing a `jsi::HostObject` with no props to
`_log` or `_toString` functions caused by accessing the last element of
an empty list.

## Test plan

```ts
runOnUI(() => {
  'worklet';
  const x = makeShareableCloneOnUIRecursive(() => {
    'worklet';
    console.log(42);
  });
  _log(x);
})();
```