Fix types in `createAnimatedComponent` (#5135)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

I've been recently refactoring `createAnimatedComponent`, but I've
forgot to update types.
* Replace any with unknown
* Type viewConfig as ViewConfig
* Add 'use strict'

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Aleksandra Cynk <aleksandra.cynk@swmansion.com>