Fix `ShadowNodeWrapper` type (#5260)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR improves typing of `ShadowNodeWrapper` so it's no longer
possible to confuse it with a regular `object`.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->