Move `setNativeProps` and `getRelativeCoords` to `platformFunctions` (#5261)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR moves `setNativeProps` (and `getRelativeCoords`) function to
`platformFunctions` directory as well as updates imports and removes
circular dependencies by splitting `utils.ts` with mostly unrelated
functions (which was the root cause of the issue) into separate files.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->