Refactor `IS_NATIVE` and `global._IS_FABRIC` (#5250)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR removes `IS_NATIVE` in favor of `SHOULD_BE_USE_WEB` to be more
consistent with `PlatformChecker`. Additionally `isFabric()` was added
to `PlatformChecker`.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->