Update first sentence tense in getting-started.mdx (#5287)
Having "was" confused me, and I went back to the fundamentals header,
thinking I missed something. This makes it more clear I am in the right
spot, and I don't need to go search for a "fundaments section".

To me, it currently reads "now that you have a strong understanding from
the fundamentals section, here is a new sections with ..."

<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Kacper Kapuściak <39658211+kacperkapusciak@users.noreply.github.com>