Remove some occurrences of `any` type (#5137)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

https://github.com/software-mansion/react-native-reanimated/pull/4421/files

## Summary
Remove some unnecessary casts to `any` and provide valid types for
`yogaValues`
We have found out that:
* most of keys in `yogaValues` are optional (only `windowWidth` and
`windowHeight` are required
* `yogaValues` may have more keys (`currentTransformMatrix`,
`targetTransformMatrix`)

So I've changed it type as follows
```diff
-   yogaValues: LayoutAnimationsValues,
+   yogaValues: Partial<SharedTransitionAnimationsValues>,
```
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan
I've logged a lot of `yogaValues` to compare its keys so I'm pretty sure
the new type is valid.

---------

Co-authored-by: Aleksandra Cynk <aleksandra.cynk@swmansion.com>
Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>