Remove `getContextContainerFromUIManager` and `FabricUtils.h` (#5272)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR removes `getContextContainerFromUIManager` function with unsafe
implementation using `reinterpret_cast` in favor of getting
`ContextContainer` directly from cloned `ShadowNode`.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->