Run listeners when calling sv.modify (#5306)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR fixes `modify` method of shared value so that it runs all
listeners after the change even if the reference to the value does not
change.

Co-authored-by: @kmagiera 
Co-authored-by: @piaskowyk 

## Test plan

ModifyExample.tsx