Add JSProps example (#5315)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR adds the example from
https://github.com/software-mansion/react-native-reanimated/pull/5009 to
the example app.


https://github.com/software-mansion/react-native-reanimated/assets/20516055/18353069-27c8-4a58-9ab0-19bbda16863e

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->