Update paths in iOS/tvOS/macOS workflows (#5319)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

I've noticed that MacOSExample CIs are started even for JS-only PRs
because `on.push.paths` was not specified.

Additionally, I fixed paths in iOS and tvOS workflows (`ios` &rarr;
`apple`).

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->