Add reduceMotion to mock (#5344)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

`reduceMotion` and the `ReduceMotion` enum were added in 3.5.


https://github.com/software-mansion/react-native-reanimated/releases/tag/3.5.0

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

`jest.mock('react-native-reanimated', () =>
require('react-native-reanimated/mock'))`

`FadeInUp.duration(500).reduceMotion(ReduceMotion.Never)`