Avoid double serialization of `__initData` (#5186)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

When serializing worklets, first we call `makeShareableCloneRecursive`
for `value.__initData` (see #3930) and then we process the rest of the
props. Currently, we don't skip `__initData` in the second phase which
leads to double conversion of this field and contradicts the whole idea.
The regression was introduced by me in #4944 which removed `delete
value.__initData;` in order to allow worklets to be copied to multiple
runtimes and not only one. This PR skips serializing `__initData` if it
has already been serialized.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->