Detect missing `'use strict';` (#5369)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR adds a CI check that ensures that all `*.{js,ts,tsx}` files
begin with `'use strict';` directive as well as fixes these files which
currently don't.

## Test plan

See if the CI is green.