Add option to omit `initData` in babel plugin (#5266)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR changes the babel plugin to remove `initData` on web. It aims to
reduce the bundle size.

❗ EDIT: I reworked this PR to instead have a more narrow configuration
option (`omitNativeOnlyData`). This will work well in conjunction with
our babel preset which will be able to configure the plugin
automatically based on the target platform.

## Test plan
Modify the `NextExample/babel.config.js` file:
```js
['react-native-reanimated/plugin', { omitWorkletInitData: true }]
```
Run `yarn build` in NextExample and verify that the difference between
reanimated and noreanimated is around 35KB.