Run worklets on separate thread (#5192)
## Summary

This PR introduces `runOnRuntime` function which lets you queue worklets
to run in a worklet runtime on a separate thread. This can be used for
instance to offload the JS thread when performing time-expensive
computations.

The function `runOnRuntime` can be called both from RN runtime as well
as from other worklet runtimes (including the UI runtime). Note that if
you call `runOnRuntime` from a worklet, you might need to enable
`processNestedWorklets` flag in Reanimated Babel plugin so the nested
worklet is also properly transformed. For now, `runOnRuntime` does not
suppport auto-workletization of its argument but most likely this
improvement will be added in a subsequent PR.

Here's how to use this API:

```ts
const runtime = createWorkletRuntime('fibonacci runtime', () => {
  'worklet';
  // initialize runtime if needed
  // this worklet is called on worklet runtime but synchronously on the same thread
  global.fib = (n: number) => { /* ... */ };
});

runOnRuntime(runtime, () => {
  'worklet';
  // this worklet is queued and run asynchronously on a separate thread
  const result = fib(40);
  runOnJS(setResult)(result);
});
```

Analogously to already existing `runOnUI` and `runOnJS` functions,
`runOnRuntime` returns a new function that accepts an arbitrary number
of arguments:
```ts
runOnRuntime(runtime, (x: number, y: string) => {
  'worklet';
  console.log(x, y);
})(42, "foo");
```

Here's the idea behind memory management of worklet runtimes and async
queue:
* `WorkletRuntime` keeps `std::shared_ptr<AsyncQueue>` which is
initialized during first call to `runOnRuntime`
* In its constructor, `AsyncQueue` spins up a new `std::thread` which
runs the run loop provided as C++ lambda.
* It was a deliberate decision not to call `thread.join()` in
`~AsyncQueue` because this could potentially delay the cleanup of other
parts of the library. Instead, we call `thread.detach()` and let the
thread run even after `AsyncQueue` instance is destroyed.
* This means that the queue may be still running even after the app
reload is complete.
* In this approach, the run loop lambda can still access task queue,
mutex and conditional variable after `AsyncQueue` is deallocated, so we
needed to extract these objects into a separate structure,
`AsyncQueueState`, which we pass as a `shared_ptr` to the lambda as well
as hold it as a field in `AsyncQueue` class. All usages of these objects
from `AsyncQueue` methods are delegated to `state_->` object.
* There's no way to terminate worklet runtime or async queue thread
while it's still performing some tasks, so the user is responsible for
keeping the tasks finite.
* `AsyncQueueState` holds a FIFO queue of tasks (as
`std::function<void()>`) which retain `WorkletRuntime`. This prevents
deallocating worklet runtime (as C++ object) when user no longer holds
it in JS code.
* Once all tasks are complete and the queue is empty, there is no strong
pointers to worklet runtime, so it can be properly deallocated along
with async queue and async queue state.

## Test plan

1. Convert `struct AsyncQueueState` into a `class`, add the following
destructor and set a breakpoint side:
```cpp
~AsyncQueueState() {
  int x = 42;
  (void)x;
}
```
2. Add destructor in `AsyncQueue` and set a breakpoint inside
3. Add destructor in `WorkletRuntime` and set a breakpoint inside
4. Launch the app
5. Reload the app (e.g. press <kbd>r</kbd> in Metro console or in iOS
simulator)
6. Confirm that the breakpoint in `~WorkletRuntime` is hit (because it
destroys the UI runtime)
7. Open WorkletRuntimeExample
8. Press "createWorkletRuntime"
9. Reload the app
10. Confirm that the breakpoint in `~WorkletRuntime` is hit twice
(because now it also destroys worklet runtime)
11. Press "runOnRuntime from JS"
12. Reload the app
10. Confirm that the breakpoint in `~WorkletRuntime` is hit twice as
well as breakpoints in `~AsyncQueue` and `~AsyncQueueState` are also hit
once