Bump react-native-screens (#5397)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

We've lost our searchbar on Android 😱 But once we bump
`react-native-screens` it works again 💪
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

|BEFORE|AFTER|
|--|--|
|![Screenshot 2023-11-09 at 10 21
15](https://github.com/software-mansion/react-native-reanimated/assets/56199675/971d8b2f-6e4c-49c7-bd2f-6b3406dc5739)|![Screenshot
2023-11-09 at 10 21
41](https://github.com/software-mansion/react-native-reanimated/assets/56199675/67ee4825-eb8b-403d-915b-a9f5efd982f2)|

## Test plan
I've tested on Android (Example and Fabric Exampe)
<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Aleksandra Cynk <aleksandracynk@Aleksandras-MacBook-Pro-3.local>