Update PlatformChecker.ts (#5440)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR fixes the following error by adding a missing cast.


![](https://github.com/software-mansion/react-native-reanimated/assets/20516055/9251987e-ccad-4a5e-b91f-eb43be068d8d)

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->