Remove dummy element on animation cancel (#5452)
## Summary

In exiting animation, we create a dummy - clone of the original element
that will be removed. When animation ends, we remove dummy from parent.
The problem is, if something cancels animation, our dummy component
won't be removed.

I've stumbled upon this problem while working on custom keyframes.

## Test plan

Tested on example app. Mostly on Keyframe animation example (on branch
with custom keyframes implementation).