Improve TS in src [1/6] - Remove unused overloads (#5482)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Currently worklets RGBtoHSV and HSVtoRGB are overloaded so they either
accept 3 arguments or 1 argument that has 3 properties. However, we only
use the first case so removing those overloads simplifies the code.

## Test plan

See that object overload is not used.

Co-authored by: @marmor157.