Add option to substitute `isWeb` and `shouldBeUseWeb` in babel plugin. (#5283)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR changes the Babel plugin to substitute `isWeb()` and
`shouldBeUseWeb()` calls with `true` so that Webpack is able to
recognize dead code.

## Test plan
Modify the NextExample/babel.config.js file:
```js
['react-native-reanimated/plugin', { omitNativeOnlyData: true, substituteWebPlatformChecks: true }]
```
Run yarn build in NextExample and verify that the difference between
reanimated and noreanimated is around 34KB.