Remove label "needs review" from bug report template (#5552)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
Remove label "needs review" from bug report template
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan
No tests.
<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Co-authored-by: Aleksandra Cynk <aleksandracynk@Aleksandras-MacBook-Pro-3.local>