Fix `useHandler` context issue (#5586)
## Summary

This PR fixes a regression introduced in #5518. `useHandler` works by
maintaining a `context` object that can be accessed and modified in
event handlers (e.g. within `useAnimatedGestureHandler`). Removing
`makeRemote` call there made it so that any changes to the `context`
object were not visible in event handlers, rendering the `context`
object useless. In this PR I wrapped the context object with
`makeShareable`, which works basically in the same way as `makeRemote`
did.

## Test plan

Check whether the `[SET] Modals` example behaves properly (i.e. if the
modal closing gestures are captured correctly).