Fix plugin sometimes not workletizing `ObjectMethod` (#5599)
## Summary

Fixes #5555.

Currently, when transforming functions with a `'worklet'` directive we
are only looking for:

1. FunctionDeclaration
```ts
function foo(){'worklet'}
```

2. FunctionExpression
```ts
const foo = function (){'worklet'}}
```

3. ArrowFunctionExpression
```ts
const foo = () => {'worklet'};
```

And not taking into consideration ObjectMethod:

4.
```ts
const foo = {
  bar() {'worklet'}
}
```

This PR fixes that and does a slight refactor of the upper control flow
and renames some functions.

This wasn't an issue before because plugins used in Expo and React
Native used to transform ObjectMethods to ObjectProperties:

```ts
const foo {
  bar: function() {'worklet'}
}
```

But it's no longer always the case.

## Test plan

Added a step to the CI if our App is actually bundling properly and
added a test.

## Note

I'm considering changing the names of relevant files afterwards, but in
this PR it would just be too messy I think.