Bump runner to macos-13 (#5221)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR bumps GitHub Actions runner image from `macos-12` to `macos-13`
(the most recent stable version).

> [!NOTE]  
> `macos-14` runner is still in beta

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->