Add `isWorklet` function (#4894)
## Summary

This PR adds `isWorklet` function which is meant to be used internally
as well as by other libraries in order to check whether the input value
is a worklet (i.e. is a function with `__workletHash` property added by
Reanimated Babel plugin).

Once worklet types are fixed, `isWorklet` should mark the input value as
worklet, i.e. we'll need to change its signature to `export function
isWorklet(value: unknown): value is WorkletType`.

EDIT from @tjzel:

In order to be able to use the new function `isWorklet` properly I had
to fix a bunch of types with regard to worklet types.

## Test plan

Check that TS CI passes and all the ifs with `.__workletHash` were
changed properly.

---------

Co-authored-by: Tomasz Żelawski <tomasz.zelawski@swmansion.com>