Move Reanimated Babel Plugin dependencies to dependencies (#5716)
## Summary

After thinking about this, there's no reason for us to keep up the weird
situation, with actually used Babel plugins in peer dependencies.

Reasons for that are:
- those plugins are supposed to perform a given transformation,
regardless of their version and whether they are proposals or not.
- React Native is currently in progress of changing those proposal
plugins to accepted ones, it would give it unnecessary maintenance to
check which are available.

## Test plan

I checked that our Example app works properly, there are no differences
in Jest tests and also Reanimated works with Expensify's app.