Replace `--frozen-lockfile` with `--immutable` in `yarn install` (#5747)
## Summary

This PR changes the `--frozen-lockfile` flag used during `yarn install`
to `--immutable` as recommended in Yarn docs:

> If the `--immutable` option is set (defaults to true on CI), Yarn will
abort with an error exit code if the lockfile was to be modified (other
paths can be added using the
[`immutablePatterns`](https://yarnpkg.com/configuration/yarnrc#immutablePatterns)
configuration setting). For backward compatibility we offer an alias
under the name of `--frozen-lockfile`, but it will be removed in a later
release.
>
> https://yarnpkg.com/cli/install

Even though `CI=true` is enabled by default in GitHub Actions runners, I
decided to leave `--immutable` flag in workflow steps anyway just to be
safe.

## Test plan

Check if CI is green.