Make `useEvent` return a different reference if event handler should be reattached (#5743)
## Summary

Currently Reanimated relies on reattaching event handlers on render.
However, we give any notice to React that it should occur, because
`useEvent` always returns the same reference.

With these changes, the reference for a given handler will change if it
had to be rebuild and components using `memo` will get re-rendered.

Fixes #5364 

## Test plan

See added `MemoExample.tsx`.