Check value callback whether it is valid or not (#5784)
TypeError: right operand of 'in' is not an object


![image](https://github.com/software-mansion/react-native-reanimated/assets/50390141/a37e35de-e8a0-49b5-ba04-75cea807be19)

TypeError: Cannot convert undefined value to object


![image](https://github.com/software-mansion/react-native-reanimated/assets/50390141/e9e10f0e-887a-4b27-a9f9-d6ce58bb3ea3)


<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Michał Bert <63123542+m-bert@users.noreply.github.com>