Add `jestUtils` to mocks (#5791)
## Summary

Currently we don't reexport jest specific function (especially
`setUpTests`) in our mock. This is due to fact that our recommended way
of initializing Reanimated for testing is to do that in `jest-setup.js`
file.

However, doing it just before the test suite (with `beforeAll` jest
function) is just as good, if even not better considering test
separation. This cannot be done if Reanimated is already mocked.

## Test plan

Add Reanimated to a new RN app, mock Reanimated and see that
`setUpTests` cannot be used in `beforeAll`. Do it using Reanimated built
from this PR and see that it works.