Allow SETs configuration to be updated on rerender (#5495)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
This PR changes the current SETs implementation to allow for:
- changes in configuration on rerender
- tracking the changes in components on the hidden screen (the target
snapshot is now created right before the hidden screen is shown)

## Additional changes
In this PR I also changed the logic behind applying transforms to views
in Snapshots on iOS. Now during the transition (so when the shared
element is removed from it's parent) we apply a transform to the shared
element, that is a combination of all the transforms that were applied
by superviews before removing this element from its parent. I decided to
do this in this PR, since those changes were closely tied to other
changes I was already making in the iOS Snapshot logic. I, however,
decided to not apply analogous changes on Android in this PR, because
due to the API differences, the logic there would be too complicated,
making this already humongous PR too big/dense.

TODO:
- [x] Android
- [x] `isInteractive` issues

## Test plan

Check the behavior of `[SET] Change theme` example (and possibly all the
other SET examples). Also on iOS check the `[SET] Nested transforms`
example.

---------

Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>