Fix example in our example app (#5783)
## Summary

The goal of this PR is to fix two of our examples - Matrix & Pendulum

### Pendulum
Allow to pass fractional numbers to input fields. Previously we were
filtering out everything except for digits, but we should allow users to
pass dot as well.

### Matrix
Make the example more descriptive, add additional information. Use the
same styling as for pednulum.

<table>
<td><b>BEFORE</b></td><td><b>AFTER</b></td></tr>
<td>


![image](https://github.com/software-mansion/react-native-reanimated/assets/56199675/2e59b3f8-eb6e-4361-b0d0-b193f35f132c)


</td><td>


![image](https://github.com/software-mansion/react-native-reanimated/assets/56199675/c51170c4-b782-4bba-ba1b-6cebf57726c7)

</td></tr>

</table>

## Test plan

---------

Co-authored-by: Latropos <aleksandracynk@Aleksandras-MacBook-Pro-3.local>