Fix dynamically added/removed animated styles on iOS/Android (#5268)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Fixes #5296.

1. Since this is potentially a JS-heavy fix, I'll gladly listen to
suggestions and receive help with how to profile it.

2. I personally dislike how those changes look like. I would much more
prefer to simplify some logic present in `createAnimatedComponent` and
its devil spawn, reducing some code, instead of building on top of it -
although I'm not sure if we need it at the moment.

Just watch the video to understand what is happening:

<table>
<tr><td>BEFORE (buggy)</td><td>AFTER (desired)</td></tr>
<tr><td>


https://github.com/software-mansion/react-native-reanimated/assets/40713406/07495711-a909-4499-89b8-a0be17f3d8ef

</td><td>


https://github.com/software-mansion/react-native-reanimated/assets/40713406/47cbace1-bf98-4011-8b4c-79bc6f6a80d4

</td></tr>
</table>



## Test plan

💣 

## Note

JS-Reanimated (Web implementation) still has those issues and I will
produce a separate solution for it in a finite amount of time.