Point metro resolver for react-native to lib
react-native-builder-bob uses standard directories
for "main", "module" and "react-native" in package.json.
These expect precompilation before release.
"react-native" usually points to the source directory.
Since in the case of react-native-reanimated
sometimes typescript is not transpiled correctly
and react-native uses javascript anyways
it should be safe to point "react-native" to the
compiled javascript sources instead of the typescript
sources (which are referenced in "source" anyways)