Add support for cloning `Error` (#5844)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR adds support for cloning JS `Error` instances from RN runtime to
UI runtime (or any worklet runtime).

The copied error inherits the original name, message as well as stack
trace.

Initially, I also wanted to implement cloning in the opposite direction
(UI &rarr; RN) but currently `makeShareableCloneOnUIRecursive` does not
support initializers and most likely will be subject to change as we
plan to unify the implementation of cloning across the runtimes.

## Test plan

ShareableExample.tsx

---------

Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>