Test two ways to use sharedValue in `useAnimatedStyle` (#5893)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
In many animations we have two ways of implementing them - we can use
withTiming either inside `useAnimatedStyle` or outside. For simplicity
I've decided to name test cases with `useAnimatedStyle` actively
updating shared value as "ACTIVE", and if from `useAnimatedStyle` we
observe it I call test case "PASSIVE". So these test cases are either
ACTIVE or PASSIVE from the point of view of the hook `useAnimatedStyle`


<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->