docs: fix grammar (#5978)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

This is a change of a single character in documentation. It just changes
"It can be ... ran" to "It can be ... run" because the former is bad
grammar.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Since this is a documentation-only change, there is no test plan.