fix: build issues on visionOS (#5980)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR fixes build issues for visionOS. The `CACurrentMediaTime` API
comes from QuartzCore which was not imported in this file.
Unfortunately, visionOS is more strict and requires explicit imports,
this shouldn't cause any issues on other platforms.

## Test plan

Build the app for visionOS

---------

Co-authored-by: Bartłomiej Błoniarz <bartlomiej.bloniarz@swmansion.com>