Fix eslint config (#5720)
## Summary

Because we use ESlint rules that require a parses, proper amendments to
the `.eslintrc.js` had to be made. Because of that we can also finally
extend main `.eslintrc.js` config in sub-directiories.

You can read more
[here](https://stackoverflow.com/questions/58510287/parseroptions-project-has-been-set-for-typescript-eslint-parser).

As a note, I recommend `eslint --print-config <file>` to debug if the
rules are actually extended etc.

## Test plan

🚀