Update useAnimatedKeyboard docs (#5866)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR updates the `useAnimatedKeyboard` docs to current doc pattern +
add more detailed remarks discovered while doing #5851. It's pointing to
#5851 and assumes changes from that PR, so **should be merged after**.

## Test plan

```
cd docs
yarn
yarn start
```

and go to useAnimatedKeyboard section

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Tomek Zawadzki <tomasz.zawadzki@swmansion.com>
Co-authored-by: Kacper Kapuściak <39658211+kacperkapusciak@users.noreply.github.com>