Change the docs to mention, that `.springify()` works with `.duration()` (#5990)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
Right now in our docs
[[LINK](https://docs.swmansion.com/react-native-reanimated/docs/layout-animations/entering-exiting-animations#modifiers)]
we say, thet `springify` doesn't work with duration, however it does. So
I've moved it from T**ime-based** modifiers section into **common**
ones.

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan

See in the recording that springified layout can have different
duration.

https://github.com/software-mansion/react-native-reanimated/assets/56199675/519eacb6-d011-4700-89fc-a62ac6b93684


<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->