Remove outdated code for unsupported React Native versions (#5979)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR removes old code for unsupported React Native versions as well
as updates the minimal supported version for old and new architecture.
Reanimated 3.9.0+ supports only React Native 0.71+ (or 0.74+ when new
architecture is enabled).

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->