Prevent crash on non-existent view updates in Android (#5767)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

I'm opening this PR to address a critical issue previously tackled by PR
#5216, which has stalled due to the original author's inactivity.

This fix is crucial for resolving significant bugs affecting Android
view property updates (referenced in issues #4505, #4231, #3860) that
impact application stability and performance.

Typical error : 
```unknown:NativeViewHierarchyManager: Unable to update properties for view tag 943
03-07 19:54:33.543 19722 19722 E unknown:NativeViewHierarchyManager: com.facebook.react.uimanager.IllegalViewOperationException: ViewManager for tag 943 could not be found.
```

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

### Before patching

https://github.com/software-mansion/react-native-reanimated/assets/40337934/a7cd7f8c-c0a3-4a84-98f7-4853886799f0

### After patching

https://github.com/software-mansion/react-native-reanimated/assets/40337934/c4dfe495-fed9-48cf-a227-e7768cec26cc

---------

Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>