fix: use proper classes for bridgeless (#5997)
## Summary

Thanks to @lukmccall we discovered that on bridgeless mode invoking
`getReactInstanceManager` creates another instance of application. This
PR makes it behave correctly there.

## Test plan

Run application and see that without this code, on bridgeless, there is
no additional button in dev menu since it is created and the button is
added in another instance.