Test predefined entering animation (#5995)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

The tests require about 1minute 40s to run.

The new tests:
* verify that all the exported predefined entering animations have
stable behaviour
* verify that they work with our most important modifiers: springify &
duration

Also I improve the way I deep copy updates

## Test plan


![image](https://github.com/software-mansion/react-native-reanimated/assets/56199675/97f97c49-4555-4db6-9374-676d86771f58)



<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->