Monorepo [1/?] - move precommit scripts to root (#5998)
## About

This pull request is a part of the effort of migrating to monorepo
structure:
- #5992
- #5993

## Summary

This particular pull requests moves the precommit husky script to the
root directory to enable it. It also extracts `lint-staged`
configuration from `packages/react-native-reanimated` to a separate file
for it to be less polluted.

## Test plan

The script should be working as expected. Try it by cloning the monorepo
and doing something bad in the code.