Monorepo [2/?] - use hoisted dependencies for Reanimated, app common code and paper example (#5999)
This pull request is a part of the effort of migrating to monorepo
structure:
- #5992
- #5993
- #5998

This particular pull request moves `app` directory to a package called
`reanimated-app` and `Example` directory to `reanimated-paper-example`
package. It also makes use of hoisted `node_modules` for those two
packages and `react-native-reanimated`.

Run `reanimated-paper-example` on both iOS and Android to see that it
works.