Monorepo [8/?] - docs (#6051)
## About

This pull request is a part of the effort of migrating to monorepo
structure:
- #5992
- #5993
- #5999
- #6008
- #6020
- #6023
- #6029

## Summary

This particular pull request takes our docs out of
`react-native-reanimated` workspace and turns it into a separate
package. However, due to some issues that seem to stem from Docusaurus
and Yarn versions, it doesn't work as a workspace. For now we will use
it just like we did before, with its own `yarn.lock` etc. until we are
able to fix that.

## Test plan

Docs should compile and work as expected.

---------

Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>