fix: getAnimatableRef on web (#6088)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
https://github.com/software-mansion/react-native-reanimated/pull/4533
added support for `getAnimatableRef`. Tried testing it for this
[issue](https://github.com/expo/expo/pull/29197) but found it was not
working for web.
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan
`getAnimatableRef` should work on web.
<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

---------

Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>