Fix `web-example`  (#6111)
## Summary

Following the transition to a monorepo structure, our web-example app
stopped working due to an asset loading issue. Turns out that the
`getAssetByID` function from `react-native-web` was being called from
the incorrect module. We believe this might have resulted in one module
blocking an asset while another module was trying to load it, causing
the application to crash.

This PR modifies the `metro.config.js` inside the `web-example` folder.
It is now configured to ignore the `@react-native` module from the root,
which allows assets to load correctly.

## Test plan

Run `web-example`.